// features/common.consents.v1/models/consents.ts
/**
 * Copyright (c) 2025, WSO2 LLC. (https://www.wso2.com).
 *
 * WSO2 LLC. licenses this file to you under the Apache License,
 * Version 2.0 (the "License"); you may not use this file except
 * in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

import { FunctionComponent, ReactElement, SVGProps } from "react";

/**
 * Consent interface.
 */
export interface ConsentInterface {
    id: string;
    name: string;
    displayName: string;
    description: string;
    version: number;
    type: string;
    elements: (ConsentPurposeElementInterface & Partial<ConsentElementInterface>)[];
    policyUrl?: string;
    isDefault?: boolean;
}

/**
 * Consent purpose element interface.
 */
export interface ConsentPurposeElementInterface {
    name: string;
    isMandatory: boolean;
}

/**
 * Consent element interface.
 */
export interface ConsentElementInterface {
    name: string;
    displayName: string;
    description: string;
    version: number;
    properties: Record<string, any>;
}

/**
 * Consent list item interface.
 */
export interface ConsentListItemInterface {
    id: string;
    name: string;
    displayName: string;
    type: string;
    isDefault?: boolean;
}

/**
 * Consent type interface.
 */
export interface ConsentTypeInterface {
    id: string;
    name: string;
}

/**
 * Enum for Consent types.
 */
export enum ConsentType {
    DATA_USAGE = "Data Usage",
    POLICY = "Policy"
}

/**
 * Interface for wizard step.
 */
export interface WizardStepInterface {
    name: string;
    icon: FunctionComponent<SVGProps<SVGSVGElement>> | ReactElement;
    title: string;
}